+++ src/dev/thetechnokid/rw/maths/Position.java
@@ -1,14 +1,14 @@
 package dev.thetechnokid.rw.maths;
 
 public class Position {
-	public int x, altitude;
+	public double x, altitude;
 	
 	public Position() {
 		x = 0;
 		altitude = 0;
 	}
 	
-	public Position(int x, int altitude) {
+	public Position(double x, double altitude) {
 		this.x = x;
 		this.altitude = altitude;
 	}
